+++ docs/changes.rst
@@ -34,6 +34,7 @@ Released on ? 2015.
 * Added quality check for AngularJS interpolation.
 * Added extensive group based ACLs.
 * Clarified terminology on strings needing review (formerly fuzzy).
+* Clarified terminology on strings needing action and not translated strings.
 
 weblate 2.4
 -----------